[REFACTOR] Use hooks for logics to implement SRP
