+++ app/components/LocationPermissionPopup.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { 
   Overlay, 
   Modal, 
@@ -7,6 +7,7 @@ import {
   ButtonContainer, 
   Button 
 } from "../../styles/LocationPermissionPopup.styles";
+import { useLocationPermission } from "../../hooks/useLocationPermission";
 
 interface LocationPermissionPopupProps {
   open: boolean;
@@ -15,39 +16,10 @@ interface LocationPermissionPopupProps {
   onDeny: () => void;
 }
 
-const LOCATION_PERMISSION_KEY = "locationPermission";
-
 const LocationPermissionPopup: React.FC<LocationPermissionPopupProps> = ({ open, onClose, onAllow, onDeny }) => {
-  const [permissionGranted, setPermissionGranted] = useState(false);
+  const { permissionGranted, allowPermission, denyPermission } = useLocationPermission(open, onAllow, onClose);
 
-  useEffect(() => {
-    if (open) {
-      const storedPermission = localStorage.getItem(LOCATION_PERMISSION_KEY);
-
-      if (storedPermission === "granted") {
-        setPermissionGranted(true);
-        onClose(); // Tutup popup
-        onAllow(); // Jika izin sudah ada, langsung jalankan onAllow
-      } else {
-        setPermissionGranted(false); // Reset permissionGranted jika tidak granted
-      }
-    }
-  }, [open, onAllow, onClose]);
-
-  const handleAllow = () => {
-    localStorage.setItem(LOCATION_PERMISSION_KEY, "granted");
-    setPermissionGranted(true);
-    onAllow(); // Panggil callback untuk halaman
-    onClose();
-  };
-
-  const handleDeny = () => {
-    setPermissionGranted(false);
-    onClose();
-    onDeny(); // Panggil callback untuk halaman
-  };
-
-  if (!open || permissionGranted) return null; // Jika sudah diizinkan, tidak tampilkan popup lagi
+  if (!open || permissionGranted) return null;
 
   return (
     <Overlay>
@@ -55,8 +27,8 @@ const LocationPermissionPopup: React.FC<LocationPermissionPopupProps> = ({ open,
         <WarningHeader>Lokasi Diperlukan</WarningHeader>
         <Message>Fitur ini memerlukan akses lokasi Anda. Izinkan akses?</Message>
         <ButtonContainer>
-          <Button onClick={handleDeny}>Tolak</Button>
-          <Button $primary onClick={handleAllow}>Lanjutkan</Button>
+          <Button onClick={() => { denyPermission(); onDeny(); }}>Tolak</Button>
+          <Button $primary onClick={allowPermission}>Lanjutkan</Button>
         </ButtonContainer>
       </Modal>
     </Overlay>
